python-service/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
numpy==1.24.3

# 如果使用OpenAI
# openai==1.3.0

# 如果使用本地模型
# transformers==4.35.0
# torch==2.1.0